LEVEL 4
The puzzle disappears into thin air, but now you have to face the bigger problem. You realize you don't have the knowledge to continue your search.
The knowledge is locked away in a folder. 
In order to find the location of the folder, solve the riddle.

" I hold your trash, but it's gone,
 You can bring back if u want to hang on.
 What am I? "

Traverse to this folder.